So yeah this is JBS3
Alright so this all started a long while ago (maybe like a year or 2) and I wanted to make my own VisualBasicScript type language sorta thing so I created JBS2 (for some reason it was already the 2nd) and used duktape for the javascript engine

fast forward to a week ago I wanted to see what I was cooking with JBS2 and I  realized that duktape doesn't even support ES6! So I knew V8 was the way to go and it was HARD

it took a few articles and blogs to finally correctly build and run V8 and you could say it was almost as hard as openCV but I don't think V8 made me as mad

also apparently v8_monolith.lib is TOO large so I have to zip it up for it to upload on github so if you want to use my precompiled libs you gotta extract v8.zip!