unexpected ViGEmBus crossover
I actually didn't really want to add ViGEmBus because I thought it was such it was too niche (but also because the guy who made it didn't realize the name was trademarked). I tried to make an extension dll that you could load to add more functions and constants to JBS but it seems impossible to do it that way unless i take v8 out of JBS3.exe and put it in its own dll. The problem with that is that I don't want to do that (but also because v8 would build lol it kept telling me that Clang was outdated or something)

Anyways I finally caved and just put it in JBS anyways (i mean it's my library anyways just shove it in with the rest of the features, plus the name isn't exactly the same lol it has Bus at the end)

other side tangents along the way involve SIMD for d2dvtablehijack as well as trying to understand how v8 calls a function so i could also use that in d2dvtablehijack

I'm just posting this commit now instead of later because I plan to take the classes in x360_emulation and put them in their own script like i did with hid_wiimote.js